split go test file
